Move Contact and Nav links to data directory.
